Move fee collector tests to their own module and file
